refactor(userspace-dp): split dispatch_tests.rs into per-concern files
The dispatch tests lived in a single ~1565-line dispatch_tests.rs loaded
via `#[cfg(test)] #[path = "dispatch_tests.rs"] mod tests;`, over the
modularity-discipline LOC threshold. Split it into a tests/ subdir with
per-concern sibling submodules, matching the #4667/#4664/#4665
convention: a tests/mod.rs holding the shared `use` header plus the
fixtures used by more than one concern, each submodule reached via
`use super::*`.

Concern files and their test counts:
  segmentation.rs      (7) TCP-segmentation / seg-miss counter + recorder
  shared_recycle.rs    (6) shared-recycle target-index resolution + drops
  enqueue_failure.rs   (6) enqueue_pending_forwards mirror / #1946 fabric
                           fail-closed / #2208 recycle trio / #4041 direct-TX
  ptb.rs               (6) #2301/#2328 egress-MTU PTB e2e + reply classify
  cos_shared_exact.rs  (5) #1598 shared-exact CoS TX-dispatch policy

Shared fixtures kept in tests/mod.rs (each used by 2+ concerns):
test_forwarding_with_egress_mtu, test_forwarding_decision_to_bound_ifindex,
test_live_forward_request_for_frame, and ingress_recycled_count (shared by
enqueue_failure + ptb). Concern-local fixtures travel with their tests.

The eight `super::`-prefixed body references
(super::cos::FORCE_ENQUEUE_ERR, super::FORCE_OVERSIZED,
super::FORCE_TUPLE_MISMATCH, super::UMEM_FRAME_SHIFT — all in the
enqueue_failure concern) stay byte-identical: tests/mod.rs's `use
super::*` re-globs those dispatch-module items into the `tests`
namespace, so `super::X` from a submodule resolves back to them. This was
confirmed with a standalone rustc repro before the split.

Pure test-code motion: each test/helper body is unchanged char-for-char.
The code-line multiset (sorted, non-comment/non-blank) has ZERO original
code lines missing; the only new lines are five `use super::*;` and five
`mod` declarations. The `#[test]` count is identical: 30 before == 30
after.

The parent dispatch/mod.rs declaration changed from
`#[path = "dispatch_tests.rs"] mod tests;` to plain `mod tests;`
(resolving to tests/mod.rs). The only other production edit is a
comment refresh on the test-only shared_recycle re-export.

Validation: cargo build clean; the full serial cargo suite
(--test-threads=1) passes 3860 / 0 failed / 2 ignored (matches the
pre-split baseline); `cargo test --release dispatch` runs 54 / 0 failed
(the 30 split tests plus 24 other dispatch-path tests).

Closes #4670.
