+++ userspace-dp/src/afxdp/tx/dispatch/tests/ptb.rs
@@ -1,0 +1,394 @@
+// #2301 / #2328 egress-MTU Packet-Too-Big (PTB) end-to-end tests through
+// `enqueue_pending_forwards`: ICMP Frag-Needed generation + oversized-original
+// drop, the in-MTU counter-factual, and generated-reply classification by the
+// PTB's own egress tuple (output-filter discard / DSCP rewrite / fail-closed
+// parse error). Local harness helpers `large_udp_v4_df_frame`,
+// `forwarding_for_ptb[_with_output_term]`, `run_ptb_dispatch[_with_forwarding]`.
+
+use super::*;
+
+// #2301: end-to-end egress-MTU PTB through `enqueue_pending_forwards`.
+//
+// Build a large UDP IPv4 (DF) frame in the ingress UMEM, point the forward
+// at egress ifindex 80 with a SMALL MTU, and add an egress entry for the
+// INGRESS interface (ifindex 11) so the reflected reply can be sourced.
+// The dispatcher must:
+//   - enqueue an ICMP Frag-Needed (type 3 code 4, next-hop MTU) back out
+//     the ingress binding,
+//   - NOT forward the oversized original to the egress binding,
+//   - recycle the ingress descriptor exactly once,
+//   - record the `egress_mtu_exceeded` exception.
+// The counter-factual (large MTU) proves the gate fires only on a real MTU
+// violation: the frame forwards normally and no PTB lands on the ingress.
+
+/// Build a large IPv4 UDP frame (Ethernet + IP + UDP + payload) with the
+/// DF bit set. `l3_payload` is the total L3 length excluding the 14-byte
+/// Ethernet header.
+fn large_udp_v4_df_frame(l3_payload: usize) -> Vec<u8> {
+    assert!(l3_payload >= 28, "need room for IP+UDP headers");
+    let mut frame = Vec::with_capacity(14 + l3_payload);
+    // L2: dst = firewall NIC, src = sender. EtherType IPv4.
+    frame.extend_from_slice(&[0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff]);
+    frame.extend_from_slice(&[0x00, 0x25, 0x90, 0x12, 0x34, 0x56]);
+    frame.extend_from_slice(&0x0800u16.to_be_bytes());
+    let l3 = frame.len();
+    frame.push(0x45);
+    frame.push(0x00);
+    frame.extend_from_slice(&(l3_payload as u16).to_be_bytes());
+    frame.extend_from_slice(&[0x00, 0x01]); // ID
+    frame.extend_from_slice(&0x4000u16.to_be_bytes()); // DF=1
+    frame.extend_from_slice(&[64, 17, 0x00, 0x00]); // TTL, proto UDP, csum
+    frame.extend_from_slice(&[198, 51, 100, 20]); // src
+    frame.extend_from_slice(&[203, 0, 113, 9]); // dst
+    let csum = crate::afxdp::tx::test_support::compute_ipv4_header_checksum(&frame[l3..l3 + 20]);
+    frame[l3 + 10..l3 + 12].copy_from_slice(&csum.to_be_bytes());
+    // UDP header + payload to fill out to l3_payload.
+    frame.extend_from_slice(&49152u16.to_be_bytes());
+    frame.extend_from_slice(&5201u16.to_be_bytes());
+    frame.extend_from_slice(&((l3_payload - 20) as u16).to_be_bytes());
+    frame.extend_from_slice(&0u16.to_be_bytes());
+    frame.resize(14 + l3_payload, 0xAB);
+    frame
+}
+
+/// Forwarding state for the PTB e2e test: egress 80 with `mtu`, plus an
+/// egress entry for the ingress interface (ifindex 11) carrying a primary
+/// v4 so the reflected error can be built.
+fn forwarding_for_ptb(mtu: usize) -> ForwardingState {
+    let mut forwarding = test_forwarding_with_egress_mtu(mtu);
+    forwarding.egress.insert(
+        11,
+        EgressInterface {
+            bind_ifindex: 11,
+            vlan_id: 0,
+            mtu: 1500,
+            src_mac: [0x02, 0xbf, 0x72, 0x16, 0x00, 0x01],
+            zone_id: TEST_TRUST_ZONE_ID,
+            redundancy_group: 0,
+            primary_v4: Some(std::net::Ipv4Addr::new(10, 0, 1, 1)),
+            primary_v6: None,
+        },
+    );
+    forwarding
+}
+
+fn run_ptb_dispatch(egress_mtu: usize) -> (Vec<BindingWorker>, DebugPollCounters, Vec<String>) {
+    let (bindings, dbg, _counters, reasons) =
+        run_ptb_dispatch_with_forwarding(forwarding_for_ptb(egress_mtu));
+    (bindings, dbg, reasons)
+}
+
+/// #2328: PTB e2e harness parameterized on the full `ForwardingState` so a
+/// test can install an output firewall filter / CoS classifier on the egress
+/// (ingress-reflected) interface (ifindex 11) and observe the generated PTB
+/// being classified by its OWN egress tuple. Returns the `BatchCounters` so
+/// the fail-on-revert tests can assert the generated-reply drop counters.
+fn run_ptb_dispatch_with_forwarding(
+    forwarding: ForwardingState,
+) -> (
+    Vec<BindingWorker>,
+    DebugPollCounters,
+    BatchCounters,
+    Vec<String>,
+) {
+    let mut bindings = vec![
+        BindingWorker::new_for_mirror_test(0, 0, 11, 0),
+        BindingWorker::new_for_mirror_test(1, 0, 22, 0),
+    ];
+    // 1600-byte L3 payload -> 1614-byte frame. Fits a 4096 UMEM frame but
+    // exceeds a 1400 egress MTU.
+    let frame = large_udp_v4_df_frame(1600);
+    unsafe { bindings[0].umem.area().slice_mut_unchecked(0, frame.len()) }
+        .expect("ingress frame")
+        .copy_from_slice(&frame);
+
+    let lookup = WorkerBindingLookup::from_bindings(&bindings);
+    let mirror_targets = MirrorTargetMap::default();
+    // UDP (not TCP) so the TCP-segmentation path is skipped and the
+    // egress-MTU decision is the only oversized handler.
+    let mut req = test_live_forward_request_for_frame(
+        frame.len(),
+        test_forwarding_decision_to_bound_ifindex(22),
+    );
+    req.meta.protocol = PROTO_UDP;
+    req.meta.l3_offset = 14;
+    req.meta.l4_offset = 34;
+    req.meta.pkt_len = (frame.len() - 14) as u16;
+    let mut pending = vec![req];
+    let mut post_recycles = Vec::new();
+    let ingress_ident = bindings[0].identity();
+    let ingress_live = &*bindings[0].live as *const BindingLiveState;
+    let local_tunnel_deliveries: Arc<ArcSwap<BTreeMap<i32, LocalTunnelDelivery>>> =
+        Arc::new(ArcSwap::from_pointee(BTreeMap::new()));
+    let recent_exceptions = Arc::new(Mutex::new(VecDeque::new()));
+    let worker_commands_by_id: BTreeMap<u32, Arc<Mutex<VecDeque<WorkerCommand>>>> = BTreeMap::new();
+    let mut dbg = DebugPollCounters::default();
+    let mut counters = BatchCounters::default();
+
+    let (left, rest) = bindings.split_at_mut(0);
+    let (ingress, right) = rest.split_first_mut().expect("ingress binding");
+    enqueue_pending_forwards(
+        left,
+        0,
+        ingress,
+        right,
+        &lookup,
+        &mirror_targets,
+        &mut pending,
+        &mut post_recycles,
+        1,
+        &forwarding,
+        &ingress_ident,
+        unsafe { &*ingress_live },
+        None,
+        &local_tunnel_deliveries,
+        &recent_exceptions,
+        &mut dbg,
+        &mut counters,
+        0,
+        &worker_commands_by_id,
+    );
+    let reasons: Vec<String> = recent_exceptions
+        .lock()
+        .expect("exceptions")
+        .iter()
+        .map(|e| e.reason.clone())
+        .collect();
+    (bindings, dbg, counters, reasons)
+}
+
+#[test]
+fn oversized_forward_emits_ptb_and_drops_original() {
+    let (bindings, _dbg, reasons) = run_ptb_dispatch(1400);
+
+    // PTB enqueued back out the INGRESS binding (slot 0).
+    let ingress_tx = &bindings[0].tx_pipeline.pending_tx_local;
+    assert_eq!(
+        ingress_tx.len(),
+        1,
+        "exactly one ICMP Frag-Needed must be enqueued on the ingress binding"
+    );
+    let reply = &ingress_tx[0];
+    assert_eq!(
+        reply.egress_ifindex, 11,
+        "PTB leaves via the ingress interface"
+    );
+    let b = &reply.bytes;
+    assert_eq!(
+        &b[0..6],
+        &[0x00, 0x25, 0x90, 0x12, 0x34, 0x56],
+        "reflected to the sender"
+    );
+    assert_eq!(&b[12..14], &[0x08, 0x00], "IPv4 reply");
+    assert_eq!(b[14], 0x45);
+    assert_eq!(b[23], PROTO_ICMP, "outer ICMP");
+    let icmp = 14 + 20;
+    assert_eq!(b[icmp], 3, "ICMP type 3");
+    assert_eq!(b[icmp + 1], 4, "code 4 (Frag Needed)");
+    assert_eq!(
+        u16::from_be_bytes([b[icmp + 6], b[icmp + 7]]),
+        1400,
+        "advertised next-hop MTU"
+    );
+
+    // The oversized original was NOT forwarded to the egress binding.
+    assert_eq!(
+        bindings[1].tx_pipeline.pending_tx_local.len(),
+        0,
+        "oversized original must not be forwarded"
+    );
+    assert_eq!(
+        bindings[1].tx_pipeline.pending_tx_prepared.len(),
+        0,
+        "oversized original must not be prepared for the egress TX ring"
+    );
+    // Ingress descriptor recycled exactly once (no leak, no double).
+    assert_eq!(ingress_recycled_count(&bindings[0]), 1);
+    assert!(
+        reasons.iter().any(|r| r == "egress_mtu_exceeded"),
+        "egress-MTU exception recorded: {reasons:?}"
+    );
+    // The original was NOT silently dropped without a signal: no
+    // oversized_forward_frame / slow-path drop was recorded.
+    assert!(
+        !reasons.iter().any(|r| r == "oversized_forward_frame"),
+        "MTU drop must not be miscounted as a descriptor-capacity overflow: {reasons:?}"
+    );
+}
+
+#[test]
+fn in_mtu_forward_emits_no_ptb_counterfactual() {
+    // Same 1614-byte frame, but a 9000 (jumbo) egress MTU. The frame now
+    // fits, so it forwards normally and NO PTB lands on the ingress. This
+    // is the fail-on-revert pin: if the decision wrongly fired on an
+    // in-MTU frame, the ingress would carry a spurious ICMP error.
+    let (bindings, _dbg, reasons) = run_ptb_dispatch(9000);
+    assert_eq!(
+        bindings[0].tx_pipeline.pending_tx_local.len(),
+        0,
+        "no PTB may be generated for an in-MTU frame"
+    );
+    assert!(
+        !reasons.iter().any(|r| r == "egress_mtu_exceeded"),
+        "no egress-MTU exception for an in-MTU frame: {reasons:?}"
+    );
+    // The frame was actually forwarded to the egress binding (slot 1),
+    // proving the in-MTU fast path is untouched.
+    let egress_tx = bindings[1].tx_pipeline.pending_tx_local.len()
+        + bindings[1].tx_pipeline.pending_tx_prepared.len();
+    assert_eq!(
+        egress_tx, 1,
+        "in-MTU frame must forward to the egress binding"
+    );
+}
+
+/// #2328: build a v4 output firewall filter on the PTB egress interface
+/// (ifindex 11 — the ingress interface the PTB is reflected back out of)
+/// carrying a single term, and splice it onto `forwarding_for_ptb`. The
+/// trigger frame is UDP, so a term keyed on `protocol icmp` fires ONLY for
+/// the GENERATED ICMP Frag-Needed reply — proving classify-by-the-PTB's-own
+/// egress tuple, not the UDP trigger tuple.
+fn forwarding_for_ptb_with_output_term(
+    mtu: usize,
+    term: crate::FirewallTermSnapshot,
+) -> ForwardingState {
+    let mut forwarding = forwarding_for_ptb(mtu);
+    forwarding.filter_state = crate::filter::parse_filter_state(
+        &[crate::FirewallFilterSnapshot {
+            name: "ptb-out".into(),
+            family: "inet".into(),
+            terms: vec![term],
+        }],
+        &[],
+        &[crate::InterfaceSnapshot {
+            name: "ge-0/0/0.0".into(),
+            ifindex: 11,
+            filter_output_v4: "ptb-out".into(),
+            ..Default::default()
+        }],
+        "",
+        "",
+    ).expect("filter state compiles");
+    forwarding.tx_selection_enabled_v4 = true;
+    forwarding
+}
+
+/// #2328: an OUTPUT firewall filter `then discard` matching `protocol icmp`
+/// on the PTB egress interface drops the GENERATED Frag-Needed reply — the
+/// PTB is now classified by its own egress tuple (#2238 contract parity with
+/// Time Exceeded / policy-reject / SYN-cookie). The drop lands on the
+/// dedicated `ptb_output_filter_drops` counter. Fail-on-revert: if the PTB
+/// enqueue reverts to the pre-#2328 unclassified `cos_queue_id: None,
+/// dscp_rewrite: None` (no `classify_generated_reply` call), the PTB would be
+/// enqueued and this assertion fails.
+#[test]
+fn ptb_dropped_by_egress_output_filter_discard() {
+    let (bindings, _dbg, counters, reasons) =
+        run_ptb_dispatch_with_forwarding(forwarding_for_ptb_with_output_term(
+            1400,
+            crate::FirewallTermSnapshot {
+                name: "drop-icmp".into(),
+                action: "discard".into(),
+                protocols: vec!["icmp".into()],
+                ..Default::default()
+            },
+        ));
+    // The generated PTB was DROPPED (not enqueued) by the output `discard`.
+    assert_eq!(
+        bindings[0].tx_pipeline.pending_tx_local.len(),
+        0,
+        "an output `then discard` (protocol icmp) on the egress interface must drop the generated PTB"
+    );
+    // Drop attributed to the dedicated PTB counter, not the parse-error one.
+    assert!(
+        counters.ptb_output_filter_drops >= 1,
+        "output-filter drop of the PTB must land on ptb_output_filter_drops"
+    );
+    assert_eq!(counters.generated_reply_classify_parse_errors, 0);
+    // The oversized original is still dropped (PMTUD), and the descriptor is
+    // recycled exactly once — no leak past the fail-closed drop.
+    assert_eq!(bindings[1].tx_pipeline.pending_tx_local.len(), 0);
+    assert_eq!(bindings[1].tx_pipeline.pending_tx_prepared.len(), 0);
+    assert_eq!(ingress_recycled_count(&bindings[0]), 1);
+    assert!(reasons.iter().any(|r| r == "egress_mtu_exceeded"));
+}
+
+/// #2328: an output filter matching the UDP TRIGGER tuple does NOT drop the
+/// generated ICMP PTB — discriminating proof the PTB is classified by its
+/// OWN (ICMP) egress tuple, not the trigger's (UDP). Sibling of the Time
+/// Exceeded `ignores_trigger_matching_output_filter` test.
+#[test]
+fn ptb_ignores_trigger_matching_output_filter() {
+    let (bindings, _dbg, counters, _reasons) =
+        run_ptb_dispatch_with_forwarding(forwarding_for_ptb_with_output_term(
+            1400,
+            crate::FirewallTermSnapshot {
+                name: "drop-udp".into(),
+                action: "discard".into(),
+                protocols: vec!["udp".into()],
+                ..Default::default()
+            },
+        ));
+    assert_eq!(
+        bindings[0].tx_pipeline.pending_tx_local.len(),
+        1,
+        "an output filter matching the UDP trigger must NOT drop the generated ICMP PTB"
+    );
+    assert_eq!(counters.ptb_output_filter_drops, 0);
+    assert_eq!(counters.generated_reply_classify_parse_errors, 0);
+}
+
+/// #2328: a forwarding-class / DSCP-rewrite output filter matching the
+/// generated PTB's ICMP tuple sets the enqueued PTB TxRequest's
+/// `dscp_rewrite` from the classifier verdict — NOT the pre-#2328
+/// hard-coded `None`. Fail-on-revert: reverting the PTB enqueue to
+/// `dscp_rewrite: None` makes the asserted rewrite disappear.
+#[test]
+fn ptb_dscp_rewrite_comes_from_classifier_not_none() {
+    let (bindings, _dbg, counters, _reasons) =
+        run_ptb_dispatch_with_forwarding(forwarding_for_ptb_with_output_term(
+            1400,
+            crate::FirewallTermSnapshot {
+                name: "mark-icmp".into(),
+                action: "accept".into(),
+                protocols: vec!["icmp".into()],
+                dscp_rewrite: Some(46),
+                ..Default::default()
+            },
+        ));
+    let ptb = &bindings[0].tx_pipeline.pending_tx_local;
+    assert_eq!(ptb.len(), 1, "PTB accepted (then accept) and enqueued");
+    assert_eq!(
+        ptb[0].dscp_rewrite,
+        Some(46),
+        "the PTB TxRequest dscp_rewrite must come from classify_generated_reply, not None"
+    );
+    assert_eq!(counters.ptb_output_filter_drops, 0);
+    assert_eq!(counters.generated_reply_classify_parse_errors, 0);
+}
+
+/// #2328 §6.2 fail-CLOSED canary: the PTB path routes a generated-reply parse
+/// failure through `classify_generated_reply`, which returns
+/// `drop: true, parse_error: true` for bytes it cannot re-parse. The
+/// dispatch wiring maps that verdict onto
+/// `generated_reply_classify_parse_errors` and drops the PTB (never leaking
+/// it past an output `discard`). This pins the shared mechanism the PTB
+/// enqueue depends on: truncated/malformed reply bytes fail closed.
+#[test]
+fn ptb_parse_failure_fails_closed_with_parse_error_verdict() {
+    let forwarding = forwarding_for_ptb(1400);
+    // Bytes too short for an L3 header — frame_l3_offset / the v4 parser
+    // reject them, exactly the canary a malformed PTB builder would hit.
+    let malformed: Vec<u8> = vec![0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff];
+    let verdict = classify_generated_reply(&forwarding, 11, &malformed, 1);
+    assert!(
+        verdict.drop,
+        "unparseable generated bytes must fail CLOSED (drop)"
+    );
+    assert!(
+        verdict.parse_error,
+        "the drop must be attributed as a parse error so the PTB path bumps generated_reply_classify_parse_errors"
+    );
+    assert_eq!(verdict.cos_queue_id, None);
+    assert_eq!(verdict.dscp_rewrite, None);
+}
