+++ userspace-dp/src/afxdp/tx/dispatch/tests/segmentation.rs
@@ -1,0 +1,221 @@
+// #1282 TCP-segmentation / segmentation-miss tests for the dispatch path:
+// `forwarded_tcp_may_need_segmentation`, the seg-miss counter
+// (`count_forwarded_tcp_segmentation_miss_if_needed`), and the
+// operator-visible seg-miss recorder (`record_forwarded_tcp_segmentation_miss`,
+// rate-capped). Local fixtures `test_decision` / `test_binding_identity`.
+
+use super::*;
+
+fn test_decision() -> SessionDecision {
+    SessionDecision {
+        resolution: ForwardingResolution {
+            disposition: ForwardingDisposition::ForwardCandidate,
+            local_ifindex: 0,
+            egress_ifindex: 80,
+            tx_ifindex: 11,
+            tunnel_endpoint_id: 0,
+            next_hop: None,
+            neighbor_mac: None,
+            src_mac: None,
+            tx_vlan_id: 80,
+        },
+        nat: NatDecision::default(),
+    }
+}
+
+fn test_binding_identity() -> BindingIdentity {
+    BindingIdentity {
+        slot: 0,
+        queue_id: 0,
+        worker_id: 0,
+        interface: Arc::<str>::from("reth1.0"),
+        ifindex: 11,
+    }
+}
+
+#[test]
+fn forwarded_tcp_may_need_segmentation_skips_mtu_sized_frame() {
+    let forwarding = test_forwarding_with_egress_mtu(1500);
+    let meta = UserspaceDpMeta {
+        addr_family: libc::AF_INET as u8,
+        protocol: PROTO_TCP,
+        l3_offset: 14,
+        ..UserspaceDpMeta::default()
+    };
+    let frame = vec![0u8; 14 + 1500];
+    assert!(!forwarded_tcp_may_need_segmentation(
+        &frame,
+        meta,
+        &test_decision(),
+        &forwarding,
+    ));
+}
+
+#[test]
+fn forwarded_tcp_may_need_segmentation_uses_frame_vlan_offset_over_stale_meta() {
+    let forwarding = test_forwarding_with_egress_mtu(1500);
+    let meta = UserspaceDpMeta {
+        addr_family: libc::AF_INET as u8,
+        protocol: PROTO_TCP,
+        // Stale metadata shape observed in #1282: the live frame is VLAN
+        // tagged, but metadata still points at a 14-byte Ethernet header.
+        l3_offset: 14,
+        ..UserspaceDpMeta::default()
+    };
+    let mut frame = vec![0u8; 18 + 1500];
+    frame[12] = 0x81;
+    frame[13] = 0x00;
+    frame[16] = 0x08;
+    frame[17] = 0x00;
+
+    assert!(!forwarded_tcp_may_need_segmentation(
+        &frame,
+        meta,
+        &test_decision(),
+        &forwarding,
+    ));
+}
+
+#[test]
+fn segmentation_miss_counter_skips_mtu_sized_vlan_frame_with_stale_meta() {
+    let forwarding = test_forwarding_with_egress_mtu(1500);
+    let meta = UserspaceDpMeta {
+        addr_family: libc::AF_INET as u8,
+        protocol: PROTO_TCP,
+        l3_offset: 14,
+        ..UserspaceDpMeta::default()
+    };
+    let mut frame = vec![0u8; 18 + 1500];
+    frame[12] = 0x81;
+    frame[13] = 0x00;
+    frame[16] = 0x08;
+    frame[17] = 0x00;
+    let tcp_segmentation_needed =
+        forwarded_tcp_may_need_segmentation(&frame, meta, &test_decision(), &forwarding);
+    let mut dbg = DebugPollCounters::default();
+
+    assert!(!count_forwarded_tcp_segmentation_miss_if_needed(
+        &mut dbg,
+        false,
+        tcp_segmentation_needed,
+    ));
+    assert_eq!(dbg.seg_needed_but_none, 0);
+}
+
+// #1282: a genuine segmentation miss must surface to operators in
+// release builds. Before the fix the only signal was the
+// `pub(in crate::afxdp)` counter `seg_needed_but_none` (never exported to
+// Go/CLI) plus an ungated `DBG SEG_MISS` eprintln. The eprintln is now
+// `debug-log`-only, so the durable signal must be the recorded exception.
+// This test recreates the failure mode: it drives the seg-miss recorder
+// and proves a `tcp_segmentation_miss` exception lands in the
+// operator-visible `recent_exceptions` buffer.
+#[test]
+fn segmentation_miss_records_operator_visible_exception() {
+    let forwarding = test_forwarding_with_egress_mtu(1500);
+    let request =
+        test_live_forward_request_for_frame(1518, test_forwarding_decision_to_bound_ifindex(11));
+    let ingress_ident = test_binding_identity();
+    let recent_exceptions = Arc::new(Mutex::new(VecDeque::new()));
+    let source_frame = vec![0u8; 1518];
+    let cap = std::cell::Cell::new(0u32);
+
+    record_forwarded_tcp_segmentation_miss(
+        &cap,
+        &recent_exceptions,
+        &ingress_ident,
+        &source_frame,
+        &request,
+        &forwarding,
+    );
+
+    let recent = recent_exceptions.lock().expect("lock");
+    assert_eq!(recent.len(), 1, "exactly one exception recorded");
+    let exc = recent.front().expect("recorded exception");
+    assert_eq!(exc.reason, "tcp_segmentation_miss");
+    assert_eq!(exc.packet_length, 1518);
+    assert_eq!(cap.get(), 1, "rate-cap counter advanced");
+}
+
+// #1282: the recorder must be rate-capped so a pathological per-packet
+// seg-miss cannot spin the `recent_exceptions` mutex on the hot path.
+// After 20 records the recorder is a no-op; the recent buffer also has
+// its own retention cap, so we assert the recorder stops incrementing the
+// cap counter and stops pushing new entries past the threshold.
+#[test]
+fn segmentation_miss_recorder_is_rate_capped() {
+    let forwarding = test_forwarding_with_egress_mtu(1500);
+    let request =
+        test_live_forward_request_for_frame(1518, test_forwarding_decision_to_bound_ifindex(11));
+    let ingress_ident = test_binding_identity();
+    let recent_exceptions = Arc::new(Mutex::new(VecDeque::new()));
+    let source_frame = vec![0u8; 1518];
+    let cap = std::cell::Cell::new(0u32);
+
+    // 25 calls; only the first 20 may record.
+    for _ in 0..25 {
+        record_forwarded_tcp_segmentation_miss(
+            &cap,
+            &recent_exceptions,
+            &ingress_ident,
+            &source_frame,
+            &request,
+            &forwarding,
+        );
+    }
+
+    assert_eq!(cap.get(), 20, "cap counter saturates at 20");
+    // The cap must also stop *exception generation*, not just the
+    // counter: prove only 20 exceptions were recorded across 25 calls,
+    // so the 5 over-cap calls never reached `record_exception` (and thus
+    // never locked the `recent_exceptions` mutex on the hot path).
+    assert_eq!(
+        recent_exceptions.lock().expect("lock").len(),
+        20,
+        "no exceptions recorded past the cap — the 5 over-cap calls never \
+         locked the recent_exceptions mutex",
+    );
+}
+
+#[test]
+fn segmentation_miss_counter_truth_table() {
+    let cases = [
+        (false, true, true, 1),
+        (true, true, false, 0),
+        (true, false, false, 0),
+        (false, false, false, 0),
+    ];
+
+    for (copied_source_frame, tcp_segmentation_needed, expected_counted, expected_counter) in cases
+    {
+        let mut dbg = DebugPollCounters::default();
+
+        assert_eq!(
+            count_forwarded_tcp_segmentation_miss_if_needed(
+                &mut dbg,
+                copied_source_frame,
+                tcp_segmentation_needed,
+            ),
+            expected_counted,
+        );
+        assert_eq!(dbg.seg_needed_but_none, expected_counter);
+    }
+}
+
+#[test]
+fn forwarded_tcp_may_need_segmentation_flags_oversized_frame() {
+    let forwarding = test_forwarding_with_egress_mtu(1500);
+    let meta = UserspaceDpMeta {
+        addr_family: libc::AF_INET as u8,
+        protocol: PROTO_TCP,
+        l3_offset: 14,
+        ..UserspaceDpMeta::default()
+    };
+    let frame = vec![0u8; 14 + 1600];
+    assert!(forwarded_tcp_may_need_segmentation(
+        &frame,
+        meta,
+        &test_decision(),
+        &forwarding,
+    ));
+}
